Fix AI profile loading and None handling
- Fixed missing UserProfile database query in chat_with_ai function
- Added proper None validation for user_id in task functions
- Fixed telegram_id field type from Integer to BigInteger in models
- Added comprehensive None checks in reminder_service
- Added validation in delegate_task and complete_task functions
- Created migration script for telegram_id type change
- Added comprehensive test suite for all functions
- All AI responses now use real user profile data instead of generic responses
- System fully protected against None value errors

+++ models.py
@@ -1,7 +1,7 @@
 import datetime
 import logging
 import enum
-from sqlalchemy import create_engine, Column, Integer, String, DateTime, Boolean, Text, ForeignKey, Enum, UniqueConstraint
+from sqlalchemy import create_engine, Column, Integer, String, DateTime, Boolean, Text, ForeignKey, Enum, UniqueConstraint, BigInteger
 from sqlalchemy.orm import declarative_base, sessionmaker, relationship
 from config import DATABASE_URL
 
@@ -19,7 +19,7 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     username = Column(String(255))
     first_name = Column(String(255))
     photo_url = Column(String(500))  # Telegram profile photo URL
